add BridgeMetadata to types
