+++ src/fed_rag/api/types.py
@@ -1,5 +1,14 @@
-"""Core types of fed-rag to include in public API."""
+"""Aux types of fed-rag to include in public API."""
 
+from fed_rag.types.knowledge_node import KnowledgeNode
 from fed_rag.types.rag import RAGConfig, RAGResponse, SourceNode
+from fed_rag.types.results import TestResult, TrainResult
 
-__all__ = ["RAGConfig", "SourceNode", "RAGResponse"]
+__all__ = [
+    "KnowledgeNode",
+    "RAGConfig",
+    "RAGResponse",
+    "SourceNode",
+    "TestResult",
+    "TrainResult",
+]
